added some checks

+++ server_services.py
@@ -1,6 +1,7 @@
 import random
 import string
 from datetime import datetime, timedelta
+from typing import Optional
 from uuid import UUID
 
 from DH_encrypt import DHEncrypt
@@ -29,7 +30,9 @@ class ServerService:
             raise Exception("Password incorrect.")
         return user
 
-    def __session_is_expired(self, session: Session) -> Session:
+    def __check_session(self, session: Optional[Session]) -> Session:
+        if session is None:
+            raise Exception("Session id is incorrect.")
         if session.expired_date < datetime.now():
             self.__db_service.delete_session(session)
             raise Exception("Session is expired. Please sign in again.")
@@ -48,15 +51,15 @@ class ServerService:
 
     def get_partial_key(self, session_id: UUID, pub_keys: dict) -> int:
         session = self.__db_service.get_session(session_id)
-        if session is None:
-            raise Exception("Session id is incorrect.")
-        session = self.__session_is_expired(session)
+        session = self.__check_session(session)
+        if session.partial_key_server:
+            return session.partial_key_server
         try:
             session.pub_key_1 = int(pub_keys["pub_key1"])
             session.pub_key_2 = int(pub_keys["pub_key2"])
             session.partial_key_client = int(pub_keys["partial_key_client"])
         except:
-            raise Exception("Required public keys are missing.")
+            raise Exception("Required public keys are missing or not integer.")
         encrypt = DHEncrypt(
             session.pub_key_1,
             session.pub_key_2,
@@ -68,7 +71,9 @@ class ServerService:
 
     def get_challenge(self, session: UUID) -> str:
         session = self.__db_service.get_session(session)
-        session = self.__session_is_expired(session)
+        session = self.__check_session(session)
+        if session.challenge:
+            return session.challenge
         challenge = generate_random_string()
         session.challenge = challenge
         self.__db_service.update_session(session)
